chore: move workflow mappers to providers
